chore: specify `type="module"` in package.json to ameliorate vite 2.7 "incorrectly packaged" warning (#55)
* chore: specify exports.import entry

* chore: set type="module" instead of exports

* fix: check if folders exist before removing

+++ src/index.ts
@@ -21,8 +21,9 @@ const mkdir = promisify(fs.mkdir);
   );
   const buildInfo: BuildIcons = JSON.parse(source);
 
-  await rmdir("types", { recursive: true });
-  await rmdir("lib", { recursive: true });
+  if (fs.existsSync("types")) await rmdir("types", { recursive: true });
+  if (fs.existsSync("lib")) await rmdir("lib", { recursive: true });
+  await mkdir("types");
   await mkdir("lib");
 
   const parser = new ComponentParser();
